Moved console log engine to different file
